# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library API",
    "version": "1.0.0",
    "description": "Kompletan CRUD API za upravljanje knjigama u biblioteci, povezan sa Postgres bazom podataka."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokalni Docker server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Vraća sve knjige iz biblioteke",
        "responses": {
          "200": {
            "description": "Uspešno osvežen spisak svih knjiga.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Greška na serveru pri komunikaciji sa bazom."
          }
        }
      },
      "post": {
        "summary": "Dodaje novu knjigu u biblioteku",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "author"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Na Drini ćuprija"
                  },
                  "author": {
                    "type": "string",
                    "example": "Ivo Andrić"
                  },
                  "year_published": {
                    "type": "integer",
                    "example": 1945
                  },
                  "available": {
                    "type": "boolean",
                    "default": true,
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Knjiga je uspešno sačuvana u bazi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Loš zahtev (fale obavezna polja title ili author)."
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Vraća specifičnu knjigu na osnovu ID-ja",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Jedinstveni ID knjige iz baze"
          }
        ],
        "responses": {
          "200": {
            "description": "Knjiga uspešno pronađena.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "Knjiga sa traženim ID-jem ne postoji u bazi."
          }
        }
      },
      "put": {
        "summary": "Ažurira podatke o postojećoj knjizi (Potpuna izmjena)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Jedinstveni ID knjige koja se menja"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Prokleta avlija"
                  },
                  "author": {
                    "type": "string",
                    "example": "Ivo Andrić"
                  },
                  "year_published": {
                    "type": "integer",
                    "example": 1954
                  },
                  "available": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Podaci o knjizi uspešno izmenjeni u bazi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "Knjiga sa tim ID-jem nije pronađena."
          }
        }
      },
      "patch": {
        "summary": "Djelomična izmjena knjige",
        "description": "Omogućuje izmjenu jednog ili više polja knjige. Sva polja koja se izostave u request body-ju ostaju netaknuta u bazi podataka.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Jedinstveni ID knjige koju krpamo (ažuriramo)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Harry Potter and the Deathly Hallows"
                  },
                  "author": {
                    "type": "string",
                    "example": "J.K. Rowling"
                  },
                  "year_published": {
                    "type": "integer",
                    "example": 2007
                  },
                  "available": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Knjiga je uspješno izmijenjena.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Loš zahtjev (poslan je prazan string za polje title ili author)."
          },
          "404": {
            "description": "Knjiga s tim ID-jem nije pronađena."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Briše knjigu iz biblioteke na osnovu ID-ja",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Jedinstveni ID knjige koju brišemo"
          }
        ],
        "responses": {
          "204": {
            "description": "Knjiga uspešno obrisana, nema sadržaja u odgovoru."
          },
          "404": {
            "description": "Knjiga sa tim ID-jem ne postoji."
          }
        }
      }
    },
    "/books/{id}/toggle-availability": {
      "patch": {
        "summary": "Preokreni dostupnost knjige",
        "description": "Automatski postavlja polje 'available' na suprotnu vrijednost (iz true u false i obrnuto) direktno u bazi podataka. Ne zahtijeva slanje request body-ja.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Jedinstveni ID knjige kojoj se mijenja status dostupnosti"
          }
        ],
        "responses": {
          "200": {
            "description": "Status dostupnosti je uspješno promijenjen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "Knjiga s tim ID-jem nije pronađena."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Automatski generisan ID od strane Postgresa (SERIAL)",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Na Drini ćuprija"
          },
          "author": {
            "type": "string",
            "example": "Ivo Andrić"
          },
          "year_published": {
            "type": "integer",
            "example": 1945
          },
          "available": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}